requirements.txt
fastapi==0.128.1
python-jose[cryptography]==3.5.0
passlib==1.7.4
bcrypt==4.0.1
python-multipart==0.0.22
pydantic==2.12.5
PyJWT==2.11.0
python-dotenv==1.2.1
stripe==13.2.0
uvicorn==0.40.0
pytest==9.0.2
httpx==0.28.1
testcontainers==4.14.1
pika==1.3.2
psycopg2-binary==2.9.11
